Fix #1623: dedup iterator/async-iterator field-hoisting and call-spill lowering
Extract the hoisted-local-to-field rewriting logic (issues #618, #641,
#655, #887) that was duplicated across FieldAccessRewriter and
FieldAccessYieldRewriter (IteratorMoveNextBodyBuilder.cs) and a third
near-identical copy in AsyncIteratorMoveNextBodyBuilder.cs's
InnerRewriter into a single shared HoistedFieldRewriter base class.
Each subclass now only implements what's actually unique to it
(yield/try/return handling for sync iterators; yield/await/return
handling for async iterators).

Delete the legacy, unreferenced IteratorMoveNextBodyBuilder.Build /
YieldReplacer local-based MoveNext path (dead code since
BuildWithFieldAccess replaced it; only call site used
BuildWithFieldAccess). This also removes the state-write disagreement
the issue called out (YieldReplacer resumed with state = -1 while the
live FieldAccessYieldRewriter path resumes with state = 0) since the
disagreeing code was never executed.

Extract the receiver+argument spill pattern shared by
SpillImportedInstanceCall, SpillUserInstanceCall,
SpillBaseInterfaceCall, and SpillBaseClassCall in
SpillSequenceSpiller.cs into one SpillReceiverAndArguments helper.

Pure refactor: lowering/emit output is unchanged. Core.Tests baseline
compare test (RefactoringBaselineTests) passes with no baseline edit;
full Core.Tests suite and Iterator/Yield-filtered Compiler.Tests pass.

Fixes #1623
