+++ src/Core/CodeAnalysis/Lowering/Iterators/IteratorMoveNextBodyBuilder.cs
@@ -237,7 +237,7 @@ public static class IteratorMoveNextBodyBuilder
                 // Rewrite the user's finally body so any references to
                 // hoisted locals resolve to fields on the state machine
                 // (Dispose has its own `this`, distinct from MoveNext's).
-                var disposeRewriter = new FieldAccessRewriter(smClass, thisParameter, hoistedFieldMap);
+                var disposeRewriter = new HoistedFieldRewriter(smClass, thisParameter, hoistedFieldMap);
                 var rewrittenFinally = disposeRewriter.RewriteStatement(tryStmt.FinallyBlock);
 
                 // try {} finally { … }
@@ -257,270 +257,10 @@ public static class IteratorMoveNextBodyBuilder
         return Lowerer.Lower(new BoundBlockStatement(null, statements.ToImmutable()));
     }
 
-    /// <summary>
-    /// Lightweight rewriter that maps hoisted local reads/writes to field
-    /// accesses on the synthesized state machine, using the supplied
-    /// <c>this</c> parameter (different per state-machine method).
-    /// </summary>
-    private sealed class FieldAccessRewriter : BoundTreeRewriter
+    private sealed class FieldAccessYieldRewriter : HoistedFieldRewriter
     {
-        private readonly StructSymbol smClass;
-        private readonly ParameterSymbol thisParameter;
-        private readonly Dictionary<VariableSymbol, FieldSymbol> fieldMap;
-
-        public FieldAccessRewriter(
-            StructSymbol smClass,
-            ParameterSymbol thisParameter,
-            Dictionary<VariableSymbol, FieldSymbol> fieldMap)
-        {
-            this.smClass = smClass;
-            this.thisParameter = thisParameter;
-            this.fieldMap = fieldMap;
-        }
-
-        protected override BoundExpression RewriteVariableExpression(BoundVariableExpression node)
-        {
-            if (this.fieldMap.TryGetValue(node.Variable, out var field))
-            {
-                return new BoundFieldAccessExpression(
-                    null,
-                    new BoundVariableExpression(null, this.thisParameter),
-                    this.smClass,
-                    field);
-            }
-
-            return node;
-        }
-
-        protected override BoundExpression RewriteAssignmentExpression(BoundAssignmentExpression node)
-        {
-            if (this.fieldMap.TryGetValue(node.Variable, out var field))
-            {
-                return new BoundFieldAssignmentExpression(
-                    null,
-                    this.thisParameter,
-                    this.smClass,
-                    field,
-                    this.RewriteExpression(node.Expression));
-            }
-
-            return base.RewriteAssignmentExpression(node);
-        }
-
-        // Issue #655: explicitly rewrite field-access expressions whose
-        // receiver is a BoundVariableExpression referencing the user-class
-        // `this` (hoisted as <>4__this). Ensures the proxy load is applied
-        // directly, protecting against regressions.
-        protected override BoundExpression RewriteFieldAccessExpression(BoundFieldAccessExpression node)
-        {
-            if (node.Receiver is BoundVariableExpression varExpr
-                && this.fieldMap.TryGetValue(varExpr.Variable, out var proxyField))
-            {
-                var rewrittenReceiver = new BoundFieldAccessExpression(
-                    null,
-                    new BoundVariableExpression(null, this.thisParameter),
-                    this.smClass,
-                    proxyField);
-                return new BoundFieldAccessExpression(null, rewrittenReceiver, node.StructType, node.Field);
-            }
-
-            return base.RewriteFieldAccessExpression(node);
-        }
-
-        // Issue #641: rewrite field assignments whose receiver is the
-        // user-class `this` (hoisted as <>4__this) to use an expression
-        // receiver so the emitter loads the proxy field.
-        protected override BoundExpression RewriteFieldAssignmentExpression(BoundFieldAssignmentExpression node)
-        {
-            var value = this.RewriteExpression(node.Value);
-            if (node.Receiver != null && this.fieldMap.TryGetValue(node.Receiver, out var proxyField))
-            {
-                var receiverExpr = new BoundFieldAccessExpression(
-                    null,
-                    new BoundVariableExpression(null, this.thisParameter),
-                    this.smClass,
-                    proxyField);
-                return BoundFieldAssignmentExpression.WithExpressionReceiver(null, receiverExpr, node.StructType, node.Field, value);
-            }
-
-            if (node.ReceiverExpression != null)
-            {
-                var receiverExpr = this.RewriteExpression(node.ReceiverExpression);
-                if (!ReferenceEquals(value, node.Value) || !ReferenceEquals(receiverExpr, node.ReceiverExpression))
-                {
-                    return BoundFieldAssignmentExpression.WithExpressionReceiver(null, receiverExpr, node.StructType, node.Field, value);
-                }
-            }
-            else if (!ReferenceEquals(value, node.Value))
-            {
-                return new BoundFieldAssignmentExpression(null, node.Receiver, node.StructType, node.Field, value);
-            }
-
-            return node;
-        }
-
-        protected override BoundStatement RewriteVariableDeclaration(BoundVariableDeclaration node)
-        {
-            if (this.fieldMap.TryGetValue(node.Variable, out var field))
-            {
-                return new BoundExpressionStatement(
-                    null,
-                    new BoundFieldAssignmentExpression(
-                        null,
-                        this.thisParameter,
-                        this.smClass,
-                        field,
-                        this.RewriteExpression(node.Initializer)));
-            }
-
-            return base.RewriteVariableDeclaration(node);
-        }
-
-        // Issue #887: an index assignment (`arr[i] = v`, `m[k] = v`) whose
-        // target temp is hoisted into a state-machine field can't reference the
-        // field through its VariableSymbol target. Switch to the expression
-        // target form reading the hoisted field (same fix as closure boxing,
-        // issue #618).
-        protected override BoundExpression RewriteIndexAssignmentExpression(BoundIndexAssignmentExpression node)
-        {
-            if (node.Target != null && this.fieldMap.TryGetValue(node.Target, out var targetField))
-            {
-                return BoundIndexAssignmentExpression.WithExpressionTarget(
-                    null,
-                    this.ReadHoistedField(targetField),
-                    this.RewriteExpression(node.Index),
-                    this.RewriteExpression(node.Value),
-                    node.Type);
-            }
-
-            return base.RewriteIndexAssignmentExpression(node);
-        }
-
-        // Issue #887: same fix for CLR-indexer writes (e.g. `dict["k"] = v` or
-        // `psi.Environment["k"] = v`) whose target temp is hoisted into a field.
-        protected override BoundExpression RewriteClrIndexAssignmentExpression(BoundClrIndexAssignmentExpression node)
-        {
-            if (node.Target != null && this.fieldMap.TryGetValue(node.Target, out var targetField))
-            {
-                return BoundClrIndexAssignmentExpression.WithExpressionTarget(
-                    null,
-                    this.ReadHoistedField(targetField),
-                    node.Indexer,
-                    this.RewriteArguments(node.Arguments),
-                    this.RewriteExpression(node.Value),
-                    node.Type);
-            }
-
-            return base.RewriteClrIndexAssignmentExpression(node);
-        }
-
-        private ImmutableArray<BoundExpression> RewriteArguments(ImmutableArray<BoundExpression> arguments)
-        {
-            var builder = ImmutableArray.CreateBuilder<BoundExpression>(arguments.Length);
-            foreach (var argument in arguments)
-            {
-                builder.Add(this.RewriteExpression(argument));
-            }
-
-            return builder.MoveToImmutable();
-        }
-
-        private BoundExpression ReadHoistedField(FieldSymbol field)
-        {
-            return new BoundFieldAccessExpression(
-                null,
-                new BoundVariableExpression(null, this.thisParameter),
-                this.smClass,
-                field);
-        }
-    }
-
-    public static IteratorMoveNextBody Build(
-        IteratorStateMachinePlan plan,
-        VariableSymbol stateLocal,
-        VariableSymbol currentLocal,
-        ParameterSymbol thisParameter)
-    {
-        // The MoveNext body is:
-        // 1. switch(state) { case 0: goto start; case 1: goto resume1; ... default: goto end; }
-        // 2. start: [user body with yields replaced by: current=x; state=K; return true; resumeK: state=0; ...]
-        // 3. end: state=-1; return false;
-
-        var statements = ImmutableArray.CreateBuilder<BoundStatement>();
-
-        // Create labels for each yield resume point
-        var yieldLabels = new Dictionary<int, BoundLabel>();
-        foreach (var kvp in plan.YieldStates.OrderBy(p => p.Value))
-        {
-            var label = new BoundLabel($"$iterResume_{kvp.Value}");
-            yieldLabels[kvp.Value] = label;
-        }
-
-        var startLabel = new BoundLabel("$iterStart");
-        var endLabel = new BoundLabel("$iterEnd");
-
-        // State dispatch: if state == 0 goto start; if state == K goto resumeK; else goto end
-        statements.Add(new BoundConditionalGotoStatement(
-            null,
-            startLabel,
-            new BoundBinaryExpression(
-                null,
-                new BoundVariableExpression(null, stateLocal),
-                BoundBinaryOperator.Bind(Syntax.SyntaxKind.EqualsEqualsToken, TypeSymbol.Int32, TypeSymbol.Int32),
-                new BoundLiteralExpression(null, 0)),
-            jumpIfTrue: true));
-
-        foreach (var kvp in plan.YieldStates.OrderBy(p => p.Value))
-        {
-            statements.Add(new BoundConditionalGotoStatement(
-                null,
-                yieldLabels[kvp.Value],
-                new BoundBinaryExpression(
-                    null,
-                    new BoundVariableExpression(null, stateLocal),
-                    BoundBinaryOperator.Bind(Syntax.SyntaxKind.EqualsEqualsToken, TypeSymbol.Int32, TypeSymbol.Int32),
-                    new BoundLiteralExpression(null, kvp.Value)),
-                jumpIfTrue: true));
-        }
-
-        // Default: goto end
-        statements.Add(new BoundGotoStatement(null, endLabel));
-
-        // start:
-        statements.Add(new BoundLabelStatement(null, startLabel));
-
-        // Rewrite the user body: replace yields with state transitions
-        var rewriter = new YieldReplacer(stateLocal, currentLocal, yieldLabels, endLabel);
-        var rewrittenBody = rewriter.RewriteStatement(plan.Body);
-
-        // Flatten the body into the statement list
-        if (rewrittenBody is BoundBlockStatement block)
-        {
-            statements.AddRange(block.Statements);
-        }
-        else
-        {
-            statements.Add(rewrittenBody);
-        }
-
-        // Fall through to end: state = -1; return false
-        statements.Add(new BoundLabelStatement(null, endLabel));
-        statements.Add(new BoundExpressionStatement(
-            null,
-            new BoundAssignmentExpression(null, stateLocal, new BoundLiteralExpression(null, -1))));
-        statements.Add(new BoundReturnStatement(null, new BoundLiteralExpression(null, false)));
-
-        var body = new BoundBlockStatement(null, statements.ToImmutable());
-        return new IteratorMoveNextBody(body, thisParameter);
-    }
-
-    private sealed class FieldAccessYieldRewriter : BoundTreeRewriter
-    {
-        private readonly StructSymbol smClass;
-        private readonly ParameterSymbol thisParameter;
         private readonly FieldSymbol stateField;
         private readonly FieldSymbol currentField;
-        private readonly Dictionary<VariableSymbol, FieldSymbol> fieldMap;
         private readonly Dictionary<int, BoundLabel> yieldLabels;
         private readonly BoundLabel endLabel;
         private readonly IteratorTryDispatchPlan tryDispatch;
@@ -536,134 +276,14 @@ public static class IteratorMoveNextBodyBuilder
             BoundLabel endLabel,
             IteratorTryDispatchPlan tryDispatch,
             IReadOnlyDictionary<BoundYieldStatement, int> yieldStateMap)
+            : base(smClass, thisParameter, fieldMap)
         {
-            this.smClass = smClass;
-            this.thisParameter = thisParameter;
             this.stateField = stateField;
             this.currentField = currentField;
-            this.fieldMap = fieldMap;
             this.yieldLabels = yieldLabels;
             this.endLabel = endLabel;
             this.tryDispatch = tryDispatch;
             this.yieldStateMap = yieldStateMap;
-        }
-
-        protected override BoundExpression RewriteVariableExpression(BoundVariableExpression node)
-        {
-            if (this.fieldMap.TryGetValue(node.Variable, out var field))
-            {
-                return this.FieldRead(field);
-            }
-
-            return node;
-        }
-
-        protected override BoundExpression RewriteAssignmentExpression(BoundAssignmentExpression node)
-        {
-            if (this.fieldMap.TryGetValue(node.Variable, out var field))
-            {
-                return this.FieldWrite(field, this.RewriteExpression(node.Expression));
-            }
-
-            return base.RewriteAssignmentExpression(node);
-        }
-
-        // Issue #655: explicitly rewrite field-access expressions whose
-        // receiver is a BoundVariableExpression referencing the user-class
-        // `this` (hoisted as <>4__this). Ensures the proxy load is applied
-        // directly, protecting against regressions.
-        protected override BoundExpression RewriteFieldAccessExpression(BoundFieldAccessExpression node)
-        {
-            if (node.Receiver is BoundVariableExpression varExpr
-                && this.fieldMap.TryGetValue(varExpr.Variable, out var proxyField))
-            {
-                var rewrittenReceiver = this.FieldRead(proxyField);
-                return new BoundFieldAccessExpression(null, rewrittenReceiver, node.StructType, node.Field);
-            }
-
-            return base.RewriteFieldAccessExpression(node);
-        }
-
-        // Issue #641: rewrite field assignments whose receiver is the
-        // user-class `this` (hoisted as <>4__this) to use an expression
-        // receiver so the emitter loads the proxy field.
-        protected override BoundExpression RewriteFieldAssignmentExpression(BoundFieldAssignmentExpression node)
-        {
-            var value = this.RewriteExpression(node.Value);
-            if (node.Receiver != null && this.fieldMap.TryGetValue(node.Receiver, out var proxyField))
-            {
-                var receiverExpr = this.FieldRead(proxyField);
-                return BoundFieldAssignmentExpression.WithExpressionReceiver(null, receiverExpr, node.StructType, node.Field, value);
-            }
-
-            if (node.ReceiverExpression != null)
-            {
-                var receiverExpr = this.RewriteExpression(node.ReceiverExpression);
-                if (!ReferenceEquals(value, node.Value) || !ReferenceEquals(receiverExpr, node.ReceiverExpression))
-                {
-                    return BoundFieldAssignmentExpression.WithExpressionReceiver(null, receiverExpr, node.StructType, node.Field, value);
-                }
-            }
-            else if (!ReferenceEquals(value, node.Value))
-            {
-                return new BoundFieldAssignmentExpression(null, node.Receiver, node.StructType, node.Field, value);
-            }
-
-            return node;
-        }
-
-        protected override BoundStatement RewriteVariableDeclaration(BoundVariableDeclaration node)
-        {
-            if (this.fieldMap.TryGetValue(node.Variable, out var field))
-            {
-                return new BoundExpressionStatement(null, this.FieldWrite(field, this.RewriteExpression(node.Initializer)));
-            }
-
-            return base.RewriteVariableDeclaration(node);
-        }
-
-        // Issue #887: an index assignment (`arr[i] = v`, `m[k] = v`) whose
-        // target temp is hoisted into a state-machine field can't reference the
-        // field through its VariableSymbol target. Switch to the expression
-        // target form reading the hoisted field (same fix as closure boxing,
-        // issue #618).
-        protected override BoundExpression RewriteIndexAssignmentExpression(BoundIndexAssignmentExpression node)
-        {
-            if (node.Target != null && this.fieldMap.TryGetValue(node.Target, out var targetField))
-            {
-                return BoundIndexAssignmentExpression.WithExpressionTarget(
-                    null,
-                    this.FieldRead(targetField),
-                    this.RewriteExpression(node.Index),
-                    this.RewriteExpression(node.Value),
-                    node.Type);
-            }
-
-            return base.RewriteIndexAssignmentExpression(node);
-        }
-
-        // Issue #887: same fix for CLR-indexer writes (e.g. `dict["k"] = v` or
-        // `psi.Environment["k"] = v`) whose target temp is hoisted into a field.
-        protected override BoundExpression RewriteClrIndexAssignmentExpression(BoundClrIndexAssignmentExpression node)
-        {
-            if (node.Target != null && this.fieldMap.TryGetValue(node.Target, out var targetField))
-            {
-                var args = ImmutableArray.CreateBuilder<BoundExpression>(node.Arguments.Length);
-                foreach (var argument in node.Arguments)
-                {
-                    args.Add(this.RewriteExpression(argument));
-                }
-
-                return BoundClrIndexAssignmentExpression.WithExpressionTarget(
-                    null,
-                    this.FieldRead(targetField),
-                    node.Indexer,
-                    args.MoveToImmutable(),
-                    this.RewriteExpression(node.Value),
-                    node.Type);
-            }
-
-            return base.RewriteClrIndexAssignmentExpression(node);
         }
 
         protected override BoundStatement RewriteYieldStatement(BoundYieldStatement node)
@@ -812,61 +432,6 @@ public static class IteratorMoveNextBodyBuilder
 
         private static BoundBlockStatement AsBlock(BoundStatement statement) =>
             statement as BoundBlockStatement ?? new BoundBlockStatement(null, ImmutableArray.Create(statement));
-
-        private BoundExpression FieldRead(FieldSymbol field) =>
-            new BoundFieldAccessExpression(null, new BoundVariableExpression(null, this.thisParameter), this.smClass, field);
-
-        private BoundExpression FieldWrite(FieldSymbol field, BoundExpression value) =>
-            new BoundFieldAssignmentExpression(null, this.thisParameter, this.smClass, field, value);
-    }
-
-    private sealed class YieldReplacer : BoundTreeRewriter
-    {
-        private readonly VariableSymbol stateLocal;
-        private readonly VariableSymbol currentLocal;
-        private readonly Dictionary<int, BoundLabel> yieldLabels;
-        private readonly BoundLabel endLabel;
-        private int yieldIndex;
-
-        public YieldReplacer(
-            VariableSymbol stateLocal,
-            VariableSymbol currentLocal,
-            Dictionary<int, BoundLabel> yieldLabels,
-            BoundLabel endLabel)
-        {
-            this.stateLocal = stateLocal;
-            this.currentLocal = currentLocal;
-            this.yieldLabels = yieldLabels;
-            this.endLabel = endLabel;
-        }
-
-        protected override BoundStatement RewriteYieldStatement(BoundYieldStatement node)
-        {
-            yieldIndex++;
-            var label = yieldLabels[yieldIndex];
-
-            // current = expr; state = K; return true; resumeK: state = -1;
-            var statements = ImmutableArray.CreateBuilder<BoundStatement>();
-            statements.Add(new BoundExpressionStatement(
-                null,
-                new BoundAssignmentExpression(null, currentLocal, node.Expression)));
-            statements.Add(new BoundExpressionStatement(
-                null,
-                new BoundAssignmentExpression(null, stateLocal, new BoundLiteralExpression(null, yieldIndex))));
-            statements.Add(new BoundReturnStatement(null, new BoundLiteralExpression(null, true)));
-            statements.Add(new BoundLabelStatement(null, label));
-            statements.Add(new BoundExpressionStatement(
-                null,
-                new BoundAssignmentExpression(null, stateLocal, new BoundLiteralExpression(null, -1))));
-
-            return new BoundBlockStatement(null, statements.ToImmutable());
-        }
-
-        protected override BoundStatement RewriteReturnStatement(BoundReturnStatement node)
-        {
-            // In an iterator, `return` (with or without value) means end iteration.
-            return new BoundGotoStatement(null, endLabel);
-        }
     }
 }
